# Experiments with Node bindings optimisation